# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Open Live API
  version: 1.0.0
  description: |
    REST API for the Open Live broadcast production platform.

    **Authentication**: Pass an API key via `Authorization: Bearer <key>` header or `?key=<key>` query parameter.
    The following paths are exempt from authentication: `/health`, `/ready`, `/api/v1/status`, `/api/v1/server-info`, `/api/v1/reconnect`.
servers:
  - url: /
security:
  - bearerAuth: []
  - apiKeyQuery: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyQuery:
      type: apiKey
      in: query
      name: key
  schemas:
...
    Graphic:
      type: object
      required: [id, name, url, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        url:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    GraphicInput:
      type: object
      required: [name, url]
      properties:
        name:
          type: string
          minLength: 1
        url:
          type: string
    GraphicPatch:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        url:
          type: string

    # ── Output ──────────────────────────────────────────────────────────────────
...
paths:
...
  /api/v1/graphics:
    get:
      summary: List graphics
      operationId: listGraphics
      tags: [Graphics]
      responses:
        '200':
          description: Array of graphics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Graphic'
    post:
      summary: Create graphic
      operationId: createGraphic
      tags: [Graphics]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GraphicInput'
      responses:
        '201':
          description: Created graphic
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Graphic'
        '400':
          description: Validation error
  /api/v1/graphics/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get graphic
      operationId: getGraphic
      tags: [Graphics]
      responses:
        '200':
          description: Graphic
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Graphic'
        '404':
          description: Not found
    patch:
      summary: Update graphic
      operationId: updateGraphic
      tags: [Graphics]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GraphicPatch'
      responses:
        '200':
          description: Updated graphic
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Graphic'
        '404':
          description: Not found
    delete:
      summary: Delete graphic
      operationId: deleteGraphic
      tags: [Graphics]
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found

  # ── Outputs ──────────────────────────────────────────────────────────────────
...
  /api/v1/productions/{id}/audio:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: List audio elements
      operationId: listAudioElements
      tags: [Audio]
      responses:
        '200':
          description: Audio elements
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    elementId:
                      type: string
                    blockId:
                      type: string
                    label:
                      type: string
                    mixerInput:
                      type: string
                      nullable: true
  /api/v1/productions/{id}/audio/{elementId}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: elementId
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get audio element properties
      operationId: getAudioElement
      tags: [Audio]
      responses:
        '200':
          description: Audio element properties
          content:
            application/json:
              schema:
                type: object
                properties:
                  element_id:
                    type: string
                  properties:
                    type: object
                    properties:
                      volume:
                        type: number
                      mute:
                        type: boolean
        '404':
          description: Not found
    patch:
      summary: Update audio element property
      operationId: updateAudioElement
      tags: [Audio]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [property, value]
              properties:
                property:
                  type: string
                  enum: [volume, mute]
                value:
                  oneOf:
                    - type: number
                      minimum: 0
                      maximum: 10
                    - type: boolean
      responses:
        '200':
          description: Updated properties
          content:
            application/json:
              schema:
                type: object
                properties:
                  element_id:
                    type: string
                  properties:
                    type: object
                    additionalProperties: true
        '404':
          description: Not found

  # Production — Stats
...